+++ packages/office-ui-fabric-react/src/components/ScrollablePane/ScrollablePane.types.ts
@@ -1,12 +1,15 @@
 import * as React from 'react';
-import { ScrollablePane } from './ScrollablePane';
+import { ScrollablePaneBase } from './ScrollablePane.base';
+import { IStyle, ITheme } from '../../Styling';
+import { IStyleFunction } from '../../Utilities';
 
 export interface IScrollablePane {
   /** Triggers a layout update for the pane. */
   forceLayoutUpdate(): void;
 }
 
-export interface IScrollablePaneProps extends React.HTMLAttributes<HTMLElement | ScrollablePane> {
+export interface IScrollablePaneProps extends React.HTMLAttributes<HTMLElement | ScrollablePaneBase> {
+  // export interface IScrollablePaneProps extends React.Props<ScrollablePaneBase> {
   /**
    * Optional callback to access the IScrollablePane interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
@@ -14,7 +17,42 @@ export interface IScrollablePaneProps extends React.HTMLAttributes<HTMLElement |
   componentRef?: (component: IScrollablePane) => void;
 
   /**
-   * Class name to apply to the root in addition to ms-ScrollablePane.
-  */
+   * Call to provide customized styling that will layer on top of the variant rules
+   */
+  getStyles?: IStyleFunction<IScrollablePaneStyleProps, IScrollablePaneStyles>;
+
+  /**
+   * Theme provided by HOC.
+   */
+  theme?: ITheme;
+
+  /**
+   * Additional css class to apply to the ScrollablePane
+   * @defaultvalue undefined
+   */
   className?: string;
+}
+
+export interface IScrollablePaneStyleProps {
+  /**
+   * Accept theme prop.
+   */
+  theme: ITheme;
+
+  /**
+   * Accept custom classNames
+   */
+  className?: string;
+
+  // Insert ScrollablePane style props below
+
+}
+
+export interface IScrollablePaneStyles {
+  /**
+   * Style for the root element.
+   */
+  root: IStyle;
+
+  // Insert ScrollablePane classNames below
 }
